Fixed outbreak spread trend y axis ticks also modularised the spread trend chart
