+++ src/components/OutbreakSpreadChart/index.js
@@ -1,0 +1,3 @@
+import drawTrendChart from "./OutbreakSpreadChart";
+
+export default drawTrendChart;
