+++ src/index.js
@@ -21,6 +21,7 @@ import drawHotspotMap from "./components/HotspotMap";
 import drawObservationTable from "./components/ObservationTable";
 import drawPrefectureTable from "./components/DistrictTable";
 import drawAgeTrendChart from "./components/AgeTrendChart";
+import drawTrendChart from "./components/OutbreakSpreadChart";
 
 // Keep reference to current chart in order to clean up when redrawing.
 let testingTrendChart = null;
@@ -32,13 +33,10 @@ mapboxgl.accessToken =
 const PREFECTURE_JSON_PATH = "static/districts.geojson";
 const JSON_PATH = "https://data.covid19kerala.info/summary/latest.json";
 const TIME_FORMAT = "YYYY-MM-DD";
-const COLOR_ACTIVE = "rgb(238,161,48)";
-const COLOR_CONFIRMED = "rgb(223,14,31)";
-const COLOR_RECOVERED = "rgb(25,118,210)";
-const COLOR_DECEASED = "rgb(55,71,79)";
 const COLOR_TESTED = "rgb(164,173,192)";
 const COLOR_TESTED_DAILY = "rgb(209,214,223)";
 const COLOR_INCREASE = "rgb(163,172,191)";
+
 const PAGE_TITLE = "COVID-19 Kerala Tracker";
 export const SUPPORTED_LANGS = LANGUAGES;
 let LANG = "en";
@@ -278,141 +276,6 @@ function drawMap() {
   });
 }
 
-function getRGBColor(color) {
-  return color
-    .substring(4, color.length - 1)
-    .replace(/ /g, "")
-    .split(",");
-}
-
-function drawTrendChart(sheetTrend) {
-  var cols = {
-    Date: ["Date"],
-    Confirmed: ["Confirmed"],
-    Active: ["Active"],
-    Critical: ["Critical"],
-    Deceased: ["Deceased"],
-    Recovered: ["Recovered"],
-    Tested: ["Tested"],
-  };
-
-  for (var i = 0; i < sheetTrend.length; i++) {
-    var row = sheetTrend[i];
-
-    /*if (i === 0) {
-      // Skip early feb data point
-      continue;
-    }*/
-
-    cols.Date.push(row.date);
-    cols.Confirmed.push(row.confirmedCumulative);
-    cols.Critical.push(row.criticalCumulative);
-    cols.Deceased.push(row.deceasedCumulative);
-    cols.Recovered.push(row.recoveredCumulative);
-    cols.Active.push(
-      row.confirmedCumulative - row.deceasedCumulative - row.recoveredCumulative
-    );
-    cols.Tested.push(row.testedCumulative);
-  }
-
-  var chart = c3.generate({
-    bindto: "#trend-chart",
-    data: {
-      x: "Date",
-      color: function (color, d) {
-        if (d && d.index === cols.Date.length - 2) {
-          let rgb = getRGBColor(color);
-          return `rgba(${rgb[0]},${rgb[1]},${rgb[2]},${0.6})`;
-        } else {
-          return color;
-        }
-      },
-      columns: [
-        cols.Date,
-        cols.Confirmed,
-        cols.Active,
-        cols.Recovered,
-        cols.Deceased,
-        //cols.Tested
-      ],
-      regions: {
-        [cols.Confirmed[0]]: [
-          { start: cols.Date[cols.Date.length - 2], style: "dashed" },
-        ],
-        [cols.Active[0]]: [
-          { start: cols.Date[cols.Date.length - 2], style: "dashed" },
-        ],
-        [cols.Recovered[0]]: [
-          { start: cols.Date[cols.Date.length - 2], style: "dashed" },
-        ],
-        [cols.Deceased[0]]: [
-          { start: cols.Date[cols.Date.length - 2], style: "dashed" },
-        ],
-        //[cols.Tested[0]]: [{'start': cols.Date[cols.Date.length-2], 'style':'dashed'}],
-      },
-      names: {
-        Date: i18next.t("Date"),
-        Confirmed: i18next.t("Confirmed"),
-        Active: i18next.t("Active"),
-        Recovered: i18next.t("Recovered"),
-        Deceased: i18next.t("Deceased"),
-      },
-    },
-    color: {
-      pattern: [COLOR_CONFIRMED, COLOR_ACTIVE, COLOR_RECOVERED, COLOR_DECEASED],
-    },
-    point: {
-      r: 3,
-    },
-    axis: {
-      x: {
-        type: "timeseries",
-        tick: {
-          format: "%b %d",
-          count: 6,
-        },
-      },
-      y: {
-        padding: {
-          bottom: 0,
-        },
-        tick: {
-          values: [0, 100, 200, 300, 400, 500, 600, 700, 800, 900, 1000, 1100],
-        },
-      },
-    },
-    tooltip: {
-      format: {
-        value: function (value, ratio, id, index) {
-          if (index && cols[id][index]) {
-            var diff = parseInt(value) - cols[id][index];
-            return `${value} (${(diff >= 0 ? "+" : "") + diff}) ${
-              index === cols.Date.length - 2
-                ? LANG === "en"
-                  ? "Provisional"
-                  : "അന്തിമമല്ല"
-                : ""
-            }`;
-          } else {
-            return value;
-          }
-        },
-      },
-    },
-    grid: {
-      x: {
-        show: true,
-      },
-      y: {
-        show: true,
-      },
-    },
-    padding: {
-      right: 24,
-    },
-  });
-}
-
 function drawPrefectureTrajectoryChart(prefectures) {
   const minimumConfirmed = 10;
   const filteredPrefectures = _.filter(prefectures, function (prefecture) {
@@ -544,7 +407,6 @@ function travelRestrictionsHelper(elementId, countries) {
 
 function drawKpis(totals, totalsDiff) {
   // Draw the KPI values
-  console.log(totals);
   function setKpi(key, value) {
     document.querySelector("#kpi-" + key + " .value").innerHTML = value;
   }
@@ -819,7 +681,7 @@ function setLang(lng) {
         drawObservationTable(ddb.underObservationData);
       }
 
-      drawTrendChart(ddb.trend);
+      drawTrendChart(ddb.trend, LANG);
       hotspotMap = drawHotspotMap(ddb.prefectures, LANG);
 
       drawPrefectureTrajectoryChart(ddb.prefectures);
@@ -906,7 +768,7 @@ function loadDataOnPage() {
       drawPageTitleCount(ddb.totals.confirmed);
       drawPrefectureTable(ddb.prefectures, ddb.totals, LANG);
       drawObservationTable(ddb.underObservationData);
-      drawTrendChart(ddb.trend);
+      drawTrendChart(ddb.trend, LANG);
       drawDailyIncreaseChart(ddb.trend);
       drawPrefectureTrajectoryChart(ddb.prefectures);
       drawAgeTrendChart(ddb.age);
